+++ packages/backend/src/modules/interop/plugins/hyperlane-hwr.ts
@@ -21,9 +21,11 @@ import {
   processLog,
 } from './hyperlane'
 import { findHyperlaneChain, HyperlaneConfig } from './hyperlane.config'
-import { getBridgeType } from './layerzero/layerzero-v2-ofts.plugin'
 import { findParsedAround, type ParsedTransferLog } from './logScan'
-import { getBestEffortTokenFrameworkBridgeType } from './tokenFrameworkBridgeTyping'
+import {
+  getBestEffortTokenFrameworkBridgeType,
+  getTokenFrameworkBridgeType,
+} from './tokenFrameworkBridgeTyping'
 import {
   createEventParser,
   createInteropEventType,
@@ -322,7 +324,7 @@ export class HyperlaneHwrPlugin implements InteropPluginResyncable {
       const dstTokenAddress = event.args.tokenAddress
       const srcWasBurned = hwrSent.args.burned
       const dstWasMinted = event.args.minted
-      const bridgeType = getBridgeType({
+      const bridgeType = getTokenFrameworkBridgeType({
         srcTokenAddress,
         dstTokenAddress,
         srcWasBurned,
