+++ packages/backend/src/modules/interop/plugins/layerzero/layerzero-v2-ofts.plugin.ts
@@ -1,13 +1,11 @@
-import {
-  Address32,
-  assert,
-  ChainSpecificAddress,
-  type KnownInteropBridgeType,
-} from '@l2beat/shared-pure'
+import { Address32, assert } from '@l2beat/shared-pure'
 import type { InteropConfigStore } from '../../engine/config/InteropConfigStore'
 import type { TokenMap } from '../../engine/match/TokenMap'
 import { findParsedAround } from '../logScan'
-import { getBestEffortTokenFrameworkBridgeType } from '../tokenFrameworkBridgeTyping'
+import {
+  getBestEffortTokenFrameworkBridgeType,
+  getTokenFrameworkBridgeType,
+} from '../tokenFrameworkBridgeTyping'
 import {
   createEventParser,
   createInteropEventType,
@@ -82,78 +80,6 @@ type NormalizedOFTSentAmounts = {
 type OFTSentTransferData = {
   address: Address32
   burned: boolean
-}
-
-export function getBridgeType({
-  srcTokenAddress,
-  dstTokenAddress,
-  srcWasBurned,
-  dstWasMinted,
-  srcChain,
-  dstChain,
-  tokenMap,
-  defaultBridgeType = 'burnAndMint',
-}: {
-  srcTokenAddress: Address32 | undefined
-  dstTokenAddress: Address32 | undefined
-  srcWasBurned: boolean | undefined
-  dstWasMinted: boolean | undefined
-  srcChain: string
-  dstChain: string
-  tokenMap: TokenMap
-  defaultBridgeType?: 'burnAndMint' | 'nonMinting' // defaults to burnAndMint, see above
-}): KnownInteropBridgeType | undefined {
-  if (
-    !srcTokenAddress ||
-    !dstTokenAddress ||
-    srcWasBurned === undefined ||
-    dstWasMinted === undefined
-  ) {
-    return
-  }
-
-  // chainspecificaddress does not support 'native' so we make do without the abstract map
-  if (
-    srcTokenAddress === Address32.NATIVE &&
-    dstTokenAddress === Address32.NATIVE
-  ) {
-    return 'nonMinting'
-  }
-  if (
-    srcTokenAddress === Address32.NATIVE ||
-    dstTokenAddress === Address32.NATIVE
-  ) {
-    return 'lockAndMint'
-  }
-
-  if (srcWasBurned && dstWasMinted) {
-    return 'burnAndMint'
-  }
-  if (srcWasBurned || dstWasMinted) {
-    return 'lockAndMint'
-  }
-
-  const srcAbstractToken = tokenMap.get(
-    ChainSpecificAddress.fromLong(
-      srcChain,
-      Address32.cropToEthereumAddress(srcTokenAddress),
-    ),
-  )
-  const dstAbstractToken = tokenMap.get(
-    ChainSpecificAddress.fromLong(
-      dstChain,
-      Address32.cropToEthereumAddress(dstTokenAddress),
-    ),
-  )
-  if (!srcAbstractToken || !dstAbstractToken) return
-  if (srcAbstractToken.issuer === null || dstAbstractToken.issuer === null) {
-    return defaultBridgeType
-  }
-
-  // lock + release
-  return srcAbstractToken.issuer === dstAbstractToken.issuer
-    ? 'nonMinting'
-    : 'lockAndMint'
 }
 
 function parseMatchingOFTSentTransfer(
@@ -385,7 +311,7 @@ export class LayerZeroV2OFTsPlugin implements InteropPlugin {
     const dstTokenAddress = oftReceivedPacketDelivered.args.dstTokenAddress
     const srcWasBurned = oftSentPacketSent.args.burned
     const dstWasMinted = oftReceivedPacketDelivered.args.minted
-    const bridgeType = getBridgeType({
+    const bridgeType = getTokenFrameworkBridgeType({
       srcTokenAddress,
       dstTokenAddress,
       srcWasBurned,
